Integration test for CVV_COLUMNS table
Kokoro build script invokes Dumper and Dumper extracts metadata to a local folder on a build worker,
the path to a folder will be stored in EXPORT_PATH env var.
After that, integration tests will be run and pointed to EXPORT_PATH to read CSV files from there.
The logic of integration tests is to make an SQL request to DB, extract the data and save them to memory,
then read CSV files, deserialize it, and then compare with the extraction from DB that the both data sets are
field-to-field equal.