Add extraction support for pid, label, and endtime for redshift-logs connector.
xid to pid is one-to-many, and pid to queryid is one-to-many, so adding pid doesn't change total number of records or grouping.

However, changed how STL_QUERY is joined with STL_QUERYTEXT and now they are joined additionally on userid, xid, pid.

Have confirmed in group discussion that redshift-logs connector is currently not used by any downstream services so the changes shouldn't cause any breakage.

Tested manually with a real Redshift instance and the extracted results are as expected.