Replace the spaces in redshift metadata connector output's headers with underscores. (#31)
The CSV loading process in the loader/transpiler library uses string names of enums for mapping CSV headers. Can't do this with header names with white spaces.
The alternative is to create additional enum-string converter for headers with spaces, but that just adds special cases and extra maps to maintain.
Simply changing the header names seems to be the better way, and no one seems to be using these extracted information right now.

Tested manually with a real Redshift instance and the extracted results are as expected.