Fix --database global flag help message.
The `--database` command-line flag depends on the connector used. It can
be a single database name or a comma-separated list of databases.
Sometimes it is used for connecting to the database (e.g. Redshift
connectors), sometimes it is used for extracting only specified
databases (e.g. the `teradata` connector). Sometimes this flag is
optional, sometimes it is required.