Implement the cloud dumper main class (#227)
The cloud dumper main class will be used to invoke the dumper from a
GCE instance. It is intended to be invoked without arguments and get
its configuratio from the instance metadata.

Instead of having the connector name separate from a separate metadata
attribute to the extractor configuration, have multiple dumper
configurations in the extractor configuration each including the name.