Ensure licenses are retrieved from POM/manifest
The CSV license report renderer from the JK1 license plugin collects
licenses for a project in a list in the following order:

1. License information from the POM files.
2. License information from manifest files.
3. Deducing the license by looking at the contents of license files.

It then uses the last entry from that list. If there is license
information from a POM or metadata file, we should rather use that.
The custom CSV reporter in the commit uses the first license information
that it finds using the above steps.

Furthermore, it allows overriding license information for a project.