Add splitting of long SQL queries in DBQLSqlTbl
The `SqlTextInfo` column of the `dbc.DBQLSqlTbl` table may need
splitting due to 64kB row size limit on Teradata 15. This change
introduces a new command-line option that allows for splitting
this column into multiple rows if the length is longer than
the specified max length.

By default, if the option is not used the column is extracted as is,
which may reach 31000 characters. With the option, the column is
split into multiple rows. The max length can be any integer
between 5000 and 30000 characters.

Example usage:

```
-Dteradata-logs.max-sql-length=7000
```