Speed up unit tests (#334)
The diagnostic query generation takes a long time, while it is not
necessary for unit tests. The query was executed only because of
failed connections to the non-existent database, which were attempted
from the real connector. Replacing the real connector with the test
connector eliminates the need to connect to the real database, so
the tasks are no longer failing.

The execution time decreased from 31s to 7s.

Command used for time measurement:
```
time ./gradlew cleanTest :dumper:app:test --no-build-cache
```