Fix queries printed in the summary
Queries from `teradata-logs` connector printed in the summary did not
exclude the missing columns. This means that the actual query executed
against the Teradata database did not contain the column in the SELECT
clause, while the query printed in the task summary contained it. This
is now fixed, so that the same query that is executed is printed in the
task summary.

Additionally, I've perfomed cleanup to reduce duplicated code and
unify how queries are printed in the task summary.