Fix cleaning up files and directories in test
The files and directories created in test were not always cleaned up due
to the `file` field being either null or not assigned the correct file
to delete. I'm adding `TopLevelTestFile` enum with all the possible
files and directories created in tests in the `MetadataDumperTest` class
and deleting them at the beginning and at the end of each test.