[b/342096412] Detect Java version in the main script
The `dwh-migration-dumper` script worked for Java version 9 and above.
This change makes it compatible with Java 8 also in Linux environments.

The `dwh-migration-dumper-java8` script is a wrapper for the old script with the same name. It is added to preserve backwards compatibility.

The Java detection is not applied for Windows environments. The bat scripts are added to preserve backwards compatibility.

The new scripts in the bin directory behave the same in the source and binary releases.