Group single-line scripts in a single text file
Some of the bash scripts are short enough to fit in a single line. It
will be easier to add new scripts if they are grouped in a single text
file. Otherwise, they work the same as the other scripts saved in
individual files, i.e. they are executed as separate bash scripts and
they produce the same format of the output.