[b/384669047] Add new connector to extract Redshift Serverless metrics (#743)
This PR introduced a new connector redshift-serverless-logs that will be used for extracting query logs and usage metrics from the Redshift Serverless. Currently, it gets data from the SYS_SERVERLESS_USAGE table only. This will be extended in the future based on the requirements.

New connector is introduced because Redshift Serverless provides the separate set of tables that contain the query logs and usage metrics. Additionally, the existing redshift-raw-logs connector fails when executed on Redshift Serverless because all source views are missing.