[b/404819770] Enable snowflake MFA caching
Enables an optional snowflake MFA caching mechanism. This simplifies the usage of the dumper for the users who decide to use the user&password auth flow with the MFA. This allows user to approve MFA request only once instead of receiving the push notification on every single SQL query.

This requires setting the authenticator property to username_password_mfa and adding JNA classes to the classpath. Changing the authenticator property does not affect the authentication if MFA is disabled for the user.