[b/440495128] Add Oracle TCO discovery script
This PR adds Oracle TCO discovery script that is an alternative method of extracting Oracle metadata and statistics for BigQuery Migration Assessment.

Script uses the same SQL files as dwh-migration-dumper, but executes them with SQL*Plus. This is helpful for the users who cannot run java application due to security or compliance reasons. Currently, it extracts only set of metadata and statistics required to TCO calculation. Later this can be extended with full version.